GatewayResolver: Do not update the /etc/resolv.conf
The only microservice which used this was proxy. From now on proxy has
its own dnsmasq which is used by eyeos-run-server and for nginx it uses
the dnsmasq microservice. (See the proxy commit logs for details).

We still need the 'gateway' to be added to the hosts, since it is used
by the nginx conf as the resolver.