Revert "Revert "GatewayResolver: Do not update the /etc/resolv.conf""
This reverts commit ea8bbe64550448fba747408bef3740fc35905a2b.

I hate reverting reverts. So ugly.

Nginx, even though explicitly told to use the gateway resolver insists
on using the local dnsmasq (specified in resolv.conf). This is because
it tries to resolve signup.service.consul (which is a special case) on
startup. We have fixed this by adding it to the hosts.