Forward SIGTERM to executed processes
It is important that when we receive a SIGTERM it is forwarded to all
processes so they can end gracefully. Docker sends a SIGTERM to the
process specified in the CMD sentence in the Dockerfile when doing
`docker stop`. If the container hasn't stopped in 10 seconds, docker
kills the container, and that is bad for databases (for example) and, in
the case of mysql, they need to do a crash recovery on the next startup,
that it is a sloooooow process.

So we better send the SIGTERM to everybody and let them finish
correctly.